chore: add partnerArtistDocumentsConnection as root field
