+++ src/schema/v2/__tests__/partner.test.js
@@ -1378,157 +1378,4 @@ describe("Partner type", () => {
       })
     })
   })
-
-  describe("#artistDocumentsConnection", () => {
-    let artistDocumentsResponse
-
-    beforeEach(() => {
-      artistDocumentsResponse = [
-        {
-          filename: "filename-one.pdf",
-          title: "File One",
-        },
-        {
-          filename: "filename-two.png",
-          title: "File Two",
-        },
-        {
-          filename: "filename-three.jpg",
-          title: "File Three",
-        },
-      ]
-
-      context = {
-        partnerLoader: () => Promise.resolve(partnerData),
-        partnerArtistDocumentsLoader: () => {
-          return Promise.resolve({
-            body: artistDocumentsResponse,
-            headers: {
-              "x-total-count": artistDocumentsResponse.length,
-            },
-          })
-        },
-      }
-    })
-
-    it("returns artist documents", async () => {
-      const query = gql`
-        {
-          partner(id: "bastian") {
-            artistDocumentsConnection(artistID: "artistID", first: 5) {
-              edges {
-                node {
-                  filename
-                  title
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const data = await runQuery(query, context)
-
-      expect(data).toEqual({
-        partner: {
-          artistDocumentsConnection: {
-            edges: [
-              {
-                node: {
-                  filename: "filename-one.pdf",
-                  title: "File One",
-                },
-              },
-              {
-                node: {
-                  filename: "filename-two.png",
-                  title: "File Two",
-                },
-              },
-              {
-                node: {
-                  filename: "filename-three.jpg",
-                  title: "File Three",
-                },
-              },
-            ],
-          },
-        },
-      })
-    })
-
-    it("returns hasNextPage=true when first is below total", async () => {
-      const query = gql`
-        {
-          partner(id: "bastian") {
-            artistDocumentsConnection(artistID: "artistID", first: 1) {
-              pageInfo {
-                hasNextPage
-              }
-            }
-          }
-        }
-      `
-
-      const data = await runQuery(query, context)
-
-      expect(data).toEqual({
-        partner: {
-          artistDocumentsConnection: {
-            pageInfo: {
-              hasNextPage: true,
-            },
-          },
-        },
-      })
-    })
-
-    it("returns hasNextPage=false when first is above total", async () => {
-      const query = gql`
-        {
-          partner(id: "bastian") {
-            artistDocumentsConnection(artistID: "artistID", first: 3) {
-              pageInfo {
-                hasNextPage
-              }
-            }
-          }
-        }
-      `
-
-      const data = await runQuery(query, context)
-
-      expect(data).toEqual({
-        partner: {
-          artistDocumentsConnection: {
-            pageInfo: {
-              hasNextPage: false,
-            },
-          },
-        },
-      })
-    })
-
-    it("loads the total count", async () => {
-      const query = gql`
-        {
-          partner(id: "bastian") {
-            artistDocumentsConnection(artistID: "artistID", first: 3) {
-              totalCount
-            }
-          }
-        }
-      `
-
-      const data = await runQuery(query, context)
-
-      expect(data).toEqual({
-        partner: {
-          artistDocumentsConnection: {
-            totalCount: 3,
-          },
-        },
-      })
-    })
-  })
 })
