+++ src/schema/v2/partnerArtistDocumentsConnection.ts
@@ -1,7 +1,18 @@
-import { GraphQLString, GraphQLObjectType, GraphQLNonNull } from "graphql"
+import {
+  GraphQLString,
+  GraphQLObjectType,
+  GraphQLNonNull,
+  GraphQLInt,
+} from "graphql"
 import { InternalIDFields } from "./object_identification"
 import { ResolverContext } from "types/graphql"
-import { connectionWithCursorInfo } from "./fields/pagination"
+import {
+  connectionWithCursorInfo,
+  paginationResolver,
+} from "./fields/pagination"
+import { GraphQLFieldConfig } from "graphql"
+import { pageable } from "relay-cursor-paging"
+import { convertConnectionArgsToGravityArgs } from "lib/helpers"
 
 export const PartnerArtistDocumentType = new GraphQLObjectType<
   any,
@@ -22,6 +33,55 @@ export const PartnerArtistDocumentType = new GraphQLObjectType<
   },
 })
 
-export const PartnerArtistDocumentsConnection = connectionWithCursorInfo({
-  nodeType: PartnerArtistDocumentType,
-})
+export const PartnerArtistDocumentsConnection: GraphQLFieldConfig<
+  void,
+  ResolverContext
+> = {
+  type: connectionWithCursorInfo({
+    nodeType: PartnerArtistDocumentType,
+  }).connectionType,
+  description:
+    "Retrieve all partner artist documents for a given partner and artist",
+  args: pageable({
+    partnerID: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The slug or ID of the Partner",
+    },
+    artistID: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The slug or ID of the Artist",
+    },
+    page: {
+      type: GraphQLInt,
+    },
+    size: {
+      type: GraphQLInt,
+    },
+  }),
+  resolve: async (_root, args, { partnerArtistDocumentsLoader }) => {
+    if (!partnerArtistDocumentsLoader) {
+      return null
+    }
+
+    const { page, size, offset } = convertConnectionArgsToGravityArgs(args)
+    const gravityOptions = {
+      size,
+      offset,
+      total_count: true,
+    }
+    const { body, headers } = await partnerArtistDocumentsLoader(
+      { artistId: args.artistID, partnerId: args.partnerID },
+      gravityOptions
+    )
+    const totalCount = parseInt(headers["x-total-count"] || "0", 10)
+
+    return paginationResolver({
+      totalCount,
+      offset,
+      page,
+      size,
+      body,
+      args,
+    })
+  },
+}
